requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-redis==6.0.0
djangorestframework==3.16.1
dotenv==0.9.9
drf-yasg==1.21.11
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
packaging==25.0
pika==1.3.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
requests==2.32.5
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
